Various minor tweaks
Saw your tweet and had a quick look through the latest changes and noticed some things being out-of-sync now.

This PR does a couple of things:
1. Bring the PHP version the PHPCompatibility standard checks against in line with the new minimum PHP version of 5.6.
2. Remove the command-line PHPCS arguments again.
3. Up the minimum PHPCS version to `3.1.1`. My bad - should have been `3.1.1` when I pulled the earlier changes. The `.`-prefix for the `phpcs.xml(.dist)` files was only introduced in this version, so if an earlier PHPCS version was being used, the config file wouldn't be picked up on automatically.
4. Very minor CS fix in one of the test files.
5. Slight tweak to the selective exclusions in the PHPCS config file.

I'm not sure why you added the command-line args again. If there was a particular reason, please let me know.

As for only examining the `./classes/src` files: except for the tiny fix I made in the `ComplexTest.php` file, the unit test files comply with PSR2 with some small exceptions as documented in the custom ruleset, so why exclude them ?