PHP8.1 Deprecation
One statement can pass a null to strtolower. This is deprecated in PHP8.1, and, when deprecated messages are enabled, causes 327 tests to error out in PhpSpreadsheet, and 598 in its own project. Use coercion to pass null string rather than null. A set of similar changes will be coming to PhpSpreadsheet this week.